+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/config/Cdoc2AuthClientConfiguration.java
@@ -1,0 +1,15 @@
+package ee.cyber.cdoc2.config;
+
+import java.util.Properties;
+
+import ee.cyber.cdoc2.exceptions.ConfigurationLoadingException;
+
+public interface Cdoc2AuthClientConfiguration {
+
+    static Cdoc2AuthClientConfiguration load(Properties properties)
+        throws ConfigurationLoadingException {
+        return Cdoc2AuthClientConfigurationProps.load(properties);
+    }
+
+    String getHostUrl();
+}