+++ cdoc2-lib/src/main/java/ee/cyber/cdoc2/config/Cdoc2ConfigurationProperties.java
@@ -115,6 +115,10 @@ public final class Cdoc2ConfigurationProperties {
     public static final String SMART_ID_CLIENT_TRUST_STORE_PWD
         = "smartid.client.ssl.trust-store-password";
 
+    /** Defines CDOC2 Authentication Server client properties file location */
+    public static final String AUTH_SERVER_PROPERTIES = "auth_server.properties";
+    public static final String AUTH_SERVER_CLIENT_HOST_URL = "auth.server.client.hostUrl";
+
     public static final String TAR_ENTRIES_THRESHOLD_PROPERTY = "ee.cyber.cdoc2.tarEntriesThreshold";
 
     public static boolean isOverWriteAllowed() {